requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.6
django-allauth==0.36.0
django-bootstrap3==10.0.1
django-heroku==0.3.1
gunicorn==19.9.0
idna==2.7
oauthlib==2.1.0
psycopg2==2.7.5
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
urllib3==1.23
whitenoise==3.3.1
